+++ visualization/app/codeCharta/core/data/data.validator.service.ts
@@ -24,17 +24,17 @@ export class DataValidatorService {
 
         if(node.children && node.children.length > 0) {
 
-            var names = {};
+            let names = {};
 
-            for(var i=0; i<node.children.length; i++){
+            for(let i=0; i<node.children.length; i++){
                 names[node.children[i].name] = true;
             }
 
             if(Object.keys(names).length !== node.children.length){
                 return false;
             } else {
-                var valid = true;
-                for(var j=0; j<node.children.length; j++){
+                let valid = true;
+                for(let j=0; j<node.children.length; j++){
                     valid = valid && this.hasUniqueChildren(node.children[j]);
                 }
                 return valid;
@@ -56,9 +56,9 @@ export class DataValidatorService {
 
         return new Promise((resolve, reject) => {
 
-                var ajv = require("ajv")();
-                var compare = ajv.compile(require("./schema.json"));
-                var valid = compare(data);
+                let ajv = require("ajv")();
+                let compare = ajv.compile(require("./schema.json"));
+                let valid = compare(data);
 
                 // TODO data.nodes[0] must be the root
                 if(!this.hasUniqueChildren(data.nodes[0])){