+++ visualization/app/codeCharta/ui/legendPanel/legendPanel.ts
@@ -1,11 +1,10 @@
 import angular from "angular";
 import "../../core/core.module";
-import "../../codeMap/codeMap";
-import "./legendPanel.css";
+import "../codeMap/codeMap";
 
 import {legendPanelComponent} from "./legendPanelComponent";
 
-angular.module("app.codeCharta.ui.legendPanel",["app.codeCharta.core.settings", "app.codeCharta.codeMap"]);
+angular.module("app.codeCharta.ui.legendPanel",["app.codeCharta.core.settings", "app.codeCharta.ui.codeMap"]);
 
 angular.module("app.codeCharta.ui.legendPanel").component(
     legendPanelComponent.selector, legendPanelComponent