+++ visualization/conf/webpack.test.config.js
@@ -16,62 +16,7 @@ module.exports = {
         filename: 'bundle.js',
         path: dist
     },
-    module: {
-        rules: [
-            {
-                test: /\.js$/,
-                exclude: /node_modules/,
-                loaders: [
-                    'ng-annotate-loader',
-                    'babel-loader'
-                ]
-            },
-            {
-                test: /\.html$/,
-                loaders: [
-                    'html-loader'
-                ]
-            },
-            {
-                test: /\.css$/,
-                use: [
-                    'style-loader',
-                    'css-loader'
-                ]
-            },
-            {
-                test: /\.(png|svg|jpg|gif)$/, use: [
-                'file-loader'
-            ]
-            },
-            {
-                test: /\.(ico)$/,
-                use: ['file-loader?name=[name].[ext]']
-            },
-            {
-                test: /\.(json)$/, use: [
-                'json-loader'
-            ]
-            },
-            {
-                test: /\.(woff|woff2|eot|ttf|otf)$/,
-                use: [
-                    'file-loader'
-                ]
-            },
-            {
-                test: /\.ts(x?)$/,
-                exclude: /node_modules/,
-                use: [
-                    'babel-loader', 'ts-loader'
-                ],
-            },
-            {
-                test: /\.glsl$/,
-                loaders: ['webpack-glsl-loader']
-            }
-        ]
-    },
+    module: require("./webpack.loaders.js"),
     plugins: [
         new webpack.ProvidePlugin({
             $: 'jquery',